+++ examples/resources/upcloud_server_firewall_ruleset/import.sh
@@ -1,2 +1,0 @@
-# Import an existing server firewall ruleset attachment using composite ID: {server_uuid}/{ruleset_uuid}
-terraform import upcloud_server_firewall_ruleset.example 00b52f84-24e8-4bd0-a5da-11404b99c4f8/19b2ebad-cf02-48af-88d2-9f1bde7fde41